serverless_rest_api/requirements.txt
requests
flask-lambda
boto3
